refactor: move Context and StackElem to submodule
